rename MapSearchQueryParamsContext -> MapSearchContext
